Final common lisp version
After making the common lisp version with generators, I realized that
all of the previous versions really obscured what was going on. This
one actually codifies what the operations are. The python one from
Norvig is very clever. The recursive common lisp one is somewhat
clever. But, clever obscures sometimes. And in this case, the least
clever version wins on being the shortest, at least in terms of the
file length.